% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARS_list_2_df.R
\name{ASR_list_2_df}
\alias{ASR_list_2_df}
\title{ARS_list_2_df converts a list of Ancestral State Reconstructions (such as the output from RASP) to a dataframe
This is used internally in the to be used with the ASR visualizer function}
\usage{
ASR_list_2_df(list, ML = FALSE)
}
\arguments{
\item{list}{a list of dataframes where each dataframe contains the states and the frequency for each node. Each element in the list should be named with the node name (e.g. node1, node2, node3) state column which has the states as alphabetical characters and no spaces (e.g. A, BC, ADBG) , and a freq column that states the probability of each state}

\item{ML}{a TRUE/FALSE stateing if you want to include a column for most likely state for each node. FALSE by default}
}
\description{
ARS_list_2_df converts a list of Ancestral State Reconstructions (such as the output from RASP) to a dataframe
This is used internally in the to be used with the ASR visualizer function
}
