% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gbif.R
\name{clean_gbif}
\alias{clean_gbif}
\title{Clean raw GBIF data}
\usage{
clean_gbif(df, filter = NULL)
}
\arguments{
\item{df}{a dataframe of raw GBIF records to filter}

\item{filter}{An optional filter by binomial name}
}
\description{
Clean raw GBIF data
}
