% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_filter.R
\name{KDE_filter}
\alias{KDE_filter}
\title{Clean records based off kernel density estimate filter}
\usage{
KDE_filter(points, unispecies = TRUE, low_r = 0.25, up_r = 0.75, rec_min = 5)
}
\arguments{
\item{points}{a dataframe of records}
}
\description{
Clean records based off kernel density estimate filter
}
