% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcreate.R
\name{mapcreate}
\alias{mapcreate}
\title{creates a simple map based on supplied country name list}
\usage{
mapcreate(country_names)
}
\arguments{
\item{country_names}{A vector of country names to use}
}
\description{
creates a simple map based on supplied country name list
}
