% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2poly.r
\name{points2Poly}
\alias{points2Poly}
\title{Create circles around points and/or alpha polygons(NEED TO TEST)}
\usage{
points2Poly(
  points,
  x = "x",
  y = "y",
  species = "species",
  map,
  alpha = FALSE,
  a = 1,
  type = "All",
  buffer_m = 1000
)
}
\arguments{
\item{points}{A dataframe containing at least the columns x, y, and sepcies}

\item{x}{the column name for the x coord}

\item{y}{the column name for the y coord}

\item{species}{the column name for the species}

\item{map}{a raster of the extent of the map}

\item{alpha}{TRUE/FALSE for if to run Alpha_Huyll procedure}

\item{a}{the amount of rounding to run alpha hull on}

\item{type}{not gonna lie. I forgot. Not needed at the moment}

\item{buffer_m}{amount of buffer in meters around each point}
}
\description{
Points as a dataframe with columns species, x, and y
map created from map function (used to crop the circles to a landmass)
}
