% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_collapse.R
\name{sf_collapse}
\alias{sf_collapse}
\title{Collapse shapefiles in an sf object based on unique values in a column}
\usage{
sf_collapse(sf, group_col)
}
\arguments{
\item{sf}{the sf object you wish to collapse}

\item{group_col}{the grouping column}
}
\description{
Collapse shapefiles in an sf object based on unique values in a column
}
