% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_gradient_fill.R
\name{sf_gradient_fill}
\alias{sf_gradient_fill}
\title{sf_gradient_fill creates a gradient fill for use in the vis_acr_shiny application}
\usage{
sf_gradient_fill(sf, grid)
}
\arguments{
\item{sf}{is the sf object}

\item{grid}{is the supplied grid over which the gradient will be created. Grids can be created with only a tiny bit of code}
}
\description{
sf_gradient_fill creates a gradient fill for use in the vis_acr_shiny application
}
