% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_pres_ab.R
\name{sf_pres_ab}
\alias{sf_pres_ab}
\title{sf_pres_ab will create a presence absence matrix given 2 sf objects}
\usage{
sf_pres_ab(x, y, x.unit, y.unit)
}
\arguments{
\item{x}{the first sf object}

\item{y}{the second sf object}

\item{x.unit}{the units of the x sf object. must be a column in the object}

\item{y.unit}{the units of the y sf object. must be a column in the object}
}
\description{
sf_pres_ab will create a presence absence matrix given 2 sf objects
}
