% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapcreate}
\alias{mapcreate}
\title{Belize',
Guatemala',
Honduras',
El Salvador',
Costa Rica',
Panama',
Nicaragua'),}
\usage{
mapcreate(country_names)
}
\description{
Belize',
Guatemala',
Honduras',
El Salvador',
Costa Rica',
Panama',
Nicaragua'),
}
