% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_RASP_dispersal.R
\name{read_RASP_dispersal}
\alias{read_RASP_dispersal}
\title{RASP DEC results}
\usage{
read_RASP_dispersal(file)
}
\arguments{
\item{file}{is the file path to the RASP output dispersal file}
}
\description{
RASP DEC results
}
